requirements/development.txt
-r base.txt

django-debug-toolbar>=4.2
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
coverage>=7.3
black>=23.0
flake8>=6.0
isort>=5.12